backend/requirements.txt
# FastAPI Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Supabase
supabase==2.22.0

# Machine Learning
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
httpx>=0.24.0
aiofiles==23.2.1
requests>=2.31.0

# Web Scraping (for Reddit and other scrapers)
praw
python-dotenv
pandas
tqdm
beautifulsoup4>=4.12.0
praw>=7.7.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# OpenAI Integration
openai>=1.0.0

# Type stubs for better IDE support
pandas-stubs>=2.0.0
types-requests>=2.31.0
